{{cookiecutter.project_slug}}/requirements/test.txt
-r base.txt
flake8==3.7.9
isort==4.3.21
pytest-aiohttp==0.3.0
